+++ phoncnic/src/main/java/com/ds/phoncnic/controller/GalleryRestController.java
@@ -1,19 +1,17 @@
-// package com.ds.phoncnic.controller;
+package com.ds.phoncnic.controller;
 
-// import org.springframework.http.ResponseEntity;
-// import org.springframework.web.bind.annotation.GetMapping;
-// import org.springframework.web.bind.annotation.RequestMapping;
-// import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-// import lombok.RequiredArgsConstructor;
-// import lombok.extern.log4j.Log4j2;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 
-// @RestController
-// // @Log4j2
-// // @RequiredArgsConstructor
-// @RequestMapping("/gallerydetails")
-// public class GalleryRestController {
+@RestController
+@Log4j2
+@RequiredArgsConstructor
+@RequestMapping("/gallerydetails")
+public class GalleryRestController {
     
 //     @GetMapping("/{gno}")
 //     public ResponseEntity<List<GalleryDTO>>getList
-// }
+}
